requirements.txt
fastapi
uvicorn
shutils
python-dotenv
